feat(data/complex/exponential): bounds on exp
Define `real.exp_bound` using `complex.exp_bound`.  Deduce numerical
bounds on `exp 1` analogous to those we have for pi.

---

This does the computation with norm_num in a very naive way.  Is it
worth doing it in a smarter way (e.g. a series of intermediate lemmas
giving successive partial sums of the series) to make the proofs
faster and so possibly allow more precise bounds to be given?

This only adds bounds with exactly 8 places after the decimal point
(the most that can be done with this naive use of norm_num without
timing out).  Should we have a series of bounds with different numbers
of decimal places, as we do for pi?

When using these bounds, it was actually `exp (-1)`, `exp (-2)` and
`exp (-3)` I wanted to bound.  Should mathlib have any bounds on `exp`
of constants other than `1`?