feat(geometry/euclidean/monge_point): orthocentric systems (#4448)
Define orthocentric systems of points, and prove some basic properties
of them.  In particular, if we say that an orthocentric system
consists of four points, one of which is the orthocenter of the
triangle formed by the other three, then each of the points is the
orthocenter of the triangle formed by the other three, and all four
triangles have the same circumradius.