refactor(linear_algebra/matrix): generalize `is_basis.to_matrix` to other index types
`is_basis.to_matrix` was defined only for families of vectors that have the
same index type as the basis, resulting in square matrices.
With some small tweaks to the definition, it's also possible to support
rectangular matrices.

This is especially helpful if you have two bases with different index sets,
and you want to apply a basis change.

Adding explicit `decidable_eq` parameters was required in some lemmas even
though `open_locale classical` was active. The right hand side of the
equalities contained `classical.decidable_eq`, which needed extra `congr`s to
unify with the actual `decidable_eq` that was available.