Redefine `cramer` to be based on `update_column` instead of `update_row`.
This really just moves a transpose operation from the definition of
`cramer_map` to the definition of `adjugate`. I claim the new definitions
have the slight advantage of being more conventional.

In the course of doing this I noticed that there still seem to be a few
places where the words 'row' and 'column' are backwards. I fixed the
instances that I encountered but did not attempt an exhaustive search.