feat(ring_theory/fractional_ideal): helper lemmas for Dedekind domains
An assortment of lemmas about `fractional_ideal`s, used in the Dedekind domain
project.

The motivation for creating this PR is that we are planning to remove the
general `has_inv` instance on `fractional_ideal` (reserving it only for
Dedekind domains), and we don't want to do the resulting refactoring twice.
So we make sure everything is in the master branch, do the refactoring there,
then merge the changes back into the work in progress branch.