Further unbundling: lie_ring_module
This was trickier to get right than it might appear, especially
to organise in such a way that `lie_ring_self_module` does not introduce
a has_bracket instance that is not defeq equal to the original. This
complication does not arise in the associative case since there we have
has_scalar and has_mul, whereas here we are using just has_bracket.

The code still needs quite a bit of work but this is definite progress.