chore(*): Make arguments to `ideal.mul_mem_{right,left}` explicit
These arguments cannot be inferred from any of the other arguments, so it seems ssensible they should be implicit.
This allows many uses of `smul_mem` to be replaced with `mul_mem_left`, which is syntactically more convenient as it does not introduce an smul.
This also removes the need for some type annotations.