Rebase upon latest mathlib, to avoid merge conflicts later on.
This is to add lemmas regarding span of submodules, and finite sets contained in them. There is also a lemma stating that maximal ideals are nontrivial if the ring is not a field. @Vierkantor helped me refine and make the proofs readable + elegant, and created the lemma `mem_span_mul_finite_of_mem_mul`.

Co-Authored-By: Ashvni <ashvni.n@gmail.com>